test(window): use html helper in window tests
