+++ packages/html/src/spreadsheet/tests/spreadsheet-dialog.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { ActionButtons } from '../../action-buttons';
-import { Button } from '../../button';
+import { Window } from '../../window';
 import { Textarea } from '../../textarea/textarea';
 
 const root = ReactDOM.createRoot(
@@ -24,53 +23,30 @@ root.render(
         <div id="test-area" className="k-d-grid k-grid-cols-1">
 
             <section>
-                <div className="k-window">
-                    <div className="k-window-titlebar">
-                        <span className="k-window-title">Insert comment</span>
-                        <div className="k-window-actions">
-                            <Button fillMode="flat" icon="close"></Button>
-                        </div>
-                    </div>
-                    <div className="k-window-content">
-                        <div className="k-form k-form-md k-form-horizontal">
-                            <div className="k-form-field">
-                                <div className="k-form-label">
-                                    <label>Comment:</label>
-                                </div>
-                                <div className="k-form-field-wrap">
-                                    <Textarea />
-                                </div>
+
+                <Window title="Insert comment" actions={[ 'close' ]} actionButtons actionButtonsAlign="end">
+                    <div className="k-form k-form-md k-form-horizontal">
+                        <div className="k-form-field">
+                            <div className="k-form-label">
+                                <label>Comment:</label>
+                            </div>
+                            <div className="k-form-field-wrap">
+                                <Textarea />
                             </div>
                         </div>
                     </div>
-                    <ActionButtons alignment="end" className="k-window-buttons">
-                        <Button>Cancel</Button>
-                        <Button themeColor="primary">Insert</Button>
-                    </ActionButtons>
-                </div>
+                </Window>
             </section>
 
             <section>
-                <div className="k-window">
-                    <div className="k-window-titlebar">
-                        <span className="k-window-title">Insert image</span>
-                        <div className="k-window-actions">
-                            <Button fillMode="flat" icon="close"></Button>
-                        </div>
+                <Window title="Insert image" actions={[ 'close' ]} actionButtons actionButtonsAlign="end">
+                    <div className="k-spreadsheet-insert-image-dialog">
+                        <label>
+                            <div>Drag an image here, or click to select</div>
+                            <input type="file" />
+                        </label>
                     </div>
-                    <div className="k-window-content">
-                        <div className="k-spreadsheet-insert-image-dialog">
-                            <label>
-                                <div>Drag an image here, or click to select</div>
-                                <input type="file" />
-                            </label>
-                        </div>
-                    </div>
-                    <ActionButtons alignment="end" className="k-window-buttons">
-                        <Button>Cancel</Button>
-                        <Button themeColor="primary">Insert</Button>
-                    </ActionButtons>
-                </div>
+                </Window>
             </section>
 
         </div>
