+++ packages/html/src/window/tests/window-action-buttons.html
@@ -1,13 +1,13 @@
 <!DOCTYPE html>
 <html lang="en" class="k-no-animations">
 <head>
-    <title>Window Angular RTL</title>
+    <title>Window Action Buttons React</title>
     <meta charset="utf-8" />
     <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css"/>
     <link rel="stylesheet" href="/packages/html/assets/styles.css"/>
     <script src="/packages/html/assets/scripts.js"></script>
 </head>
 <body id="app" class="k-body">
-    <script src="/packages/html/dist/window/tests/window-angular-rtl.js"></script>
+    <script src="/packages/html/dist/window/tests/window-action-buttons.js"></script>
 </body>
 </html>
