--- tests/window/window-angular-rtl.html
@@ -1,86 +1,0 @@
-<!doctype html>
-<html lang="en" class="k-no-animations" style="--kendo-scrollbar-width:15px;">
-<head>
-    <title>Window Angular RTL</title>
-    <meta charset="utf-8">
-    <link rel="stylesheet" data-role="kendo-theme" href="/packages/default/dist/all.css">
-    <link rel="stylesheet" href="/packages/html/assets/styles.css">
-    <script src="/packages/html/assets/scripts.js"></script>
-</head>
-<body id="app" class="k-body">
-    <style>
-        .k-window {
-            width: 100%;
-            position: relative;
-        }
-
-    </style>
-    <div id="test-area" class="k-d-grid k-grid-cols-2">
-        <section>
-            <div class="k-widget k-window" dir="rtl">
-                <div class="k-window-titlebar k-dialog-titlebar">
-                    <div class="k-window-title">Angular Window</div>
-                    <div class="k-window-actions">
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-minimize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-maximize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-close"></span>
-                        </button>
-                    </div>
-                </div>
-                <div class="k-window-content">Angular window content</div>
-            </div>
-        </section>
-        <section>
-            <div class="k-widget k-window" dir="rtl">
-                <div class="k-window-titlebar k-dialog-titlebar">
-                    <div class="k-window-title"></div>
-                    <div class="k-window-actions">
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-minimize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-maximize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-close"></span>
-                        </button>
-                    </div>
-                </div>
-                <div class="k-window-content">Angular window with no title specified</div>
-            </div>
-        </section>
-        <section>
-            <div class="k-widget k-window" dir="rtl">
-                <div class="k-window-titlebar k-dialog-titlebar">
-                    <div class="k-window-title"></div>
-                </div>
-                <div class="k-window-content">Angular window with empty title and no actions</div>
-            </div>
-        </section>
-        <section>
-            <div class="k-widget k-window k-window-minimized" dir="rtl">
-                <div class="k-window-titlebar k-dialog-titlebar">
-                    <div class="k-window-title">Angular window minimized</div>
-                    <div class="k-window-actions">
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-minimize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-window-maximize"></span>
-                        </button>
-                        <button class="k-button k-button-md k-button-flat k-button-flat-base k-rounded-md k-icon-button">
-                            <span class="k-button-icon k-icon k-i-close"></span>
-                        </button>
-                    </div>
-                </div>
-                <div class="k-window-content" hidden="">Angular minimized content</div>
-            </div>
-        </section>
-    </div>
-</body>
-</html>
